Structure improvements

+++ src/CleanArchitecture.Application/Features/Career/Commands/DeleteCareer/DeleteCareerCommandHandler.cs
@@ -1,7 +1,7 @@
 ﻿using CleanArchitecture.Application.Common.App;
 using CleanArchitecture.Application.Common.ErrorMessage;
 using CleanArchitecture.Application.Common.Exceptions;
-using CleanArchitecture.Application.Messaging.Command;
+using CleanArchitecture.Application.Common.Messaging.Command;
 using CleanArchitecture.Application.Repositories.Career;
 
 namespace CleanArchitecture.Application.Features.Career.Commands.DeleteCareer
@@ -15,6 +15,9 @@ namespace CleanArchitecture.Application.Features.Career.Commands.DeleteCareer
         #region Handle
         public async Task<ApplicationResult> Handle(DeleteCareerCommand command, CancellationToken cancellationToken)
         {
+            if (command.Id < 1)
+                throw new BadRequestException(DefaultErrorMessages.CAREER_ID_IS_REQUIRED);
+
             _ = await _careerRepository.GetCareerByIdAsync(command.Id) ?? throw new NotFoundException(DefaultErrorMessages.CAREER_NOT_FOUND);
 
             var response = await _careerRepository.DeleteCareerAsync(command.Id);
